refactor(timeline): extract caption dialog state + transcribe controller from TimelineItem
Pull the caption/transcription surface (~165 lines) out of the
TimelineItem god component into a dedicated hook + child component:

- use-caption-dialog-state.ts owns the dialog state machine, transcript
  selectors, embedded-subtitle extraction, and per-cue consolidation.
- transcribe-dialog-controller.tsx renders the TranscribeDialog from
  the hook output and is memo-gated by canManageCaptions.

Host file shrinks from 3,794 to 3,588 lines and the caption logic is
now isolated from the rest of the clip-render path.
